Extract base class for links to images
