+++ api/src/main/java/com/creatubbles/api/model/ImageLinks.java
@@ -1,21 +1,18 @@
-package com.creatubbles.api.model.creation;
+package com.creatubbles.api.model;
 
 import android.support.annotation.Nullable;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class Links {
+public class ImageLinks {
 
     private String original;
 
-    @JsonProperty("full_view")
-    private String fullView;
+    @JsonProperty("list_view")
+    private String listView;
 
     @JsonProperty("list_view_retina")
     private String listViewRetina;
-
-    @JsonProperty("list_view")
-    private String listView;
 
     @JsonProperty("matrix_view_retina")
     private String matrixViewRetina;
@@ -23,25 +20,12 @@ public class Links {
     @JsonProperty("matrix_view")
     private String matrixView;
 
-    @JsonProperty("gallery_mobile")
-    private String galleryMobile;
-
     @JsonProperty("explore_mobile")
     private String exploreMobile;
-
-    private String share;
 
     @Nullable
     public String getOriginal() {
         return original;
-    }
-
-    /**
-     * @return url to image suitable for fullscreen view, original aspect ratio
-     */
-    @Nullable
-    public String getFullView() {
-        return fullView;
     }
 
     /**
@@ -68,6 +52,7 @@ public class Links {
         return matrixViewRetina;
     }
 
+
     /**
      * @return url to image in 105x105 (might be cropped)
      */
@@ -76,13 +61,6 @@ public class Links {
         return matrixView;
     }
 
-    /**
-     * @return url to image in max 300x600 size
-     */
-    @Nullable
-    public String getGalleryMobile() {
-        return galleryMobile;
-    }
 
     /**
      * @return url to image in 90x90 (might be cropped)
@@ -92,26 +70,15 @@ public class Links {
         return exploreMobile;
     }
 
-    /**
-     * @return url to image in 600x600 (might be cropped)
-     */
-    @Nullable
-    public String getShare() {
-        return share;
-    }
-
     @Override
     public String toString() {
-        return "Links{" +
+        return "ImageLinks{" +
                 "original='" + original + '\'' +
-                ", fullView='" + fullView + '\'' +
-                ", listViewRetina='" + listViewRetina + '\'' +
                 ", listView='" + listView + '\'' +
+                ", listViewRetina='" + listViewRetina + '\'' +
                 ", matrixViewRetina='" + matrixViewRetina + '\'' +
                 ", matrixView='" + matrixView + '\'' +
-                ", galleryMobile='" + galleryMobile + '\'' +
                 ", exploreMobile='" + exploreMobile + '\'' +
-                ", share='" + share + '\'' +
                 '}';
     }
 }
